refactor: relocate shell and panel harnesses outside support tree (#4117)
Move app_shell_harness, min_viewport_harness, units_panel_test_shared,
and military/naval/civilian/diplomacy panel supports to app/test/ with
shrink-only re-export stubs. Ratchet appTestSupportLocCeiling 2261 → 962.

Refs #4117
